Version 2 of Hello World
I noticed that when I open the executable file of the original Hello World program, it closes right away. With research, I found out that to fix this problem, I need to add 
system("pause");
at the end of the code so that the person who is using the program can choose when he wants to close it instead of the program itself closing automatically. I am using Windows, for Mac or Linux users, they need to write
system("read");
I found this information in the following link: https://stackoverflow.com/questions/1449324/how-to-simulate-press-any-key-to-continue